Adding support for running DHE separately.
The DHE host shouldn't be part of the swarm cluster. You'd be one
  docker ps -aq | xargs docker rm -f
away from wiping out your demo environment.

This change updates the role logic to allow for hosts to have subsets of
roles and still keep the first host builds logic.

We also offload the computation of who is a swarm host to script that
actually parses the stevedore.yaml file.