Adding a submodule for notary
We need to build notary containers at build time so we can get the right
TLS certs and configs into them. But we want to make sure that we aren't
broken by upstream notary development. So we'll use a submodule pinned
at specific commit and move that commit as it make sense.